refactor dev models
refactor dev models so they are easier to
spawn.

Also added function to convert joint transforms to world frame
